Add symlink from CATS in GOPATH to CATS in cf-release
We normally develop out of the cf-acceptance-tests that
are vendored in with cf-release on CAPI and our aliases
assume we are using that one.

This will hopefully fix an issue where fasd_cd (z command)
takes us into the GOPATH one without realizing.

Eventually when we no longer use cf-release this can be
removed and we should work out of the GOPATH.